+++ src/smbase/typ.h
@@ -5,6 +5,8 @@
 #ifndef __TYP_H
 #define __TYP_H
 
+#include <stdint.h>
+
 // byte
 typedef unsigned char byte;
 typedef signed char signed_byte;
@@ -30,8 +32,8 @@ typedef signed char signed_byte;
 
 // This used when I want to cast a pointer to an integer for something
 // like hashing the address.  It need not be injective.
-inline long pointerToInteger(void const *p)
-  { return (long)p; }
+inline intptr_t pointerToInteger(void const *p)
+  { return (intptr_t)p; }
 
 
 // This can be used to compare two pointers, even when they do not point
@@ -45,7 +47,7 @@ inline int comparePointerAddresses(void const *p, void const *q)
   // existence of this function at least ensures I only have to change
   // one place.
   return p==q?               0 :
-         (long)p < (long)q? -1 :      // would use std::less<> here
+         (intptr_t)p < (intptr_t)q? -1 :      // would use std::less<> here
                             +1 ;
 }
